cleaned example + refactor
