+++ bioptim/examples/stochastic_optimal_control/arm_reaching_torque_driven.py
@@ -43,16 +43,16 @@ from bioptim import (
 def get_force_field(q, force_field_magnitude):
     l1 = 0.3
     l2 = 0.33
-    F_forceField = force_field_magnitude * (l1 * cas.cos(q[0]) + l2 * cas.cos(q[0] + q[1]))
+    f_force_field = force_field_magnitude * (l1 * cas.cos(q[0]) + l2 * cas.cos(q[0] + q[1]))
     hand_pos = cas.MX(2, 1)
     hand_pos[0] = l2 * cas.sin(q[0] + q[1]) + l1 * cas.sin(q[0])
     hand_pos[1] = l2 * cas.sin(q[0] + q[1])
-    tau_force_field = -F_forceField @ hand_pos
+    tau_force_field = -f_force_field @ hand_pos
     return tau_force_field
 
 
-def get_excitation_with_feedback(K, EE, EE_ref, wS):
-    return K @ ((EE - EE_ref) + wS)
+def get_excitation_with_feedback(k, hand_pos_velo, ref, sensory_noise):
+    return k @ ((hand_pos_velo - ref) + sensory_noise)
 
 
 def stochastic_forward_dynamics(
@@ -61,8 +61,8 @@ def stochastic_forward_dynamics(
     parameters: cas.MX | cas.SX,
     stochastic_variables: cas.MX | cas.SX,
     nlp: NonLinearProgram,
-    wM,
-    wS,
+    motor_noise,
+    sensory_noise,
     force_field_magnitude,
     with_gains,
 ) -> DynamicsEvaluation:
@@ -71,31 +71,29 @@ def stochastic_forward_dynamics(
     qddot = DynamicsFunctions.get(nlp.states["qddot"], states)
     qdddot = DynamicsFunctions.get(nlp.controls["qdddot"], controls)
     tau = DynamicsFunctions.get(nlp.controls["tau"], controls)
-    n_states = nlp.states.cx_start.shape[0]
-    n_controls = nlp.controls.cx_start.shape[0]
     n_q = q.shape[0]
     n_qdot = qdot.shape[0]
     n_tau = tau.shape[0]
 
     tau_fb = tau
     if with_gains:
-        ee_ref = DynamicsFunctions.get(nlp.stochastic_variables["ee_ref"], stochastic_variables)
+        ref = DynamicsFunctions.get(nlp.stochastic_variables["ref"], stochastic_variables)
         k = DynamicsFunctions.get(nlp.stochastic_variables["k"], stochastic_variables)
-        K_matrix = cas.MX(n_q + n_qdot, n_tau)
+        k_matrix = cas.MX(n_q + n_qdot, n_tau)
         for s0 in range(n_q + n_qdot):
             for s1 in range(n_tau):
-                K_matrix[s0, s1] = k[s0 * n_tau + s1]
-        K_matrix = K_matrix.T
+                k_matrix[s0, s1] = k[s0 * n_tau + s1]
+        k_matrix = k_matrix.T
 
         hand_pos = nlp.model.markers(q)[2][:2]
         hand_vel = nlp.model.marker_velocities(q, qdot)[2][:2]
         ee = cas.vertcat(hand_pos, hand_vel)
 
-        tau_fb += get_excitation_with_feedback(K_matrix, ee, ee_ref, wS)
+        tau_fb += get_excitation_with_feedback(k_matrix, ee, ref, sensory_noise)
 
     tau_force_field = get_force_field(q, force_field_magnitude)
 
-    torques_computed = tau_fb + wM + tau_force_field
+    torques_computed = tau_fb + motor_noise + tau_force_field
     dq_computed = qdot
 
     friction = np.array([[0.05, 0.025], [0.025, 0.05]])
@@ -113,7 +111,7 @@ def stochastic_forward_dynamics(
     return DynamicsEvaluation(dxdt=cas.vertcat(dq_computed, dqdot_computed, dqddot_computed), defects=defects)
 
 
-def configure_stochastic_optimal_control_problem(ocp: OptimalControlProgram, nlp: NonLinearProgram, wM, wS):
+def configure_stochastic_optimal_control_problem(ocp: OptimalControlProgram, nlp: NonLinearProgram, motor_noise, sensory_noise):
     ConfigureProblem.configure_q(ocp, nlp, True, False, False)
     ConfigureProblem.configure_qdot(ocp, nlp, True, False, True)
     ConfigureProblem.configure_qddot(ocp, nlp, True, False, True)
@@ -122,7 +120,7 @@ def configure_stochastic_optimal_control_problem(ocp: OptimalControlProgram, nlp
 
     # Stochastic variables
     ConfigureProblem.configure_stochastic_k(ocp, nlp, n_noised_controls=2, n_feedbacks=4)
-    ConfigureProblem.configure_stochastic_ee_ref(ocp, nlp, n_references=4)
+    ConfigureProblem.configure_stochastic_ref(ocp, nlp, n_references=4)
     ConfigureProblem.configure_stochastic_m(ocp, nlp, n_noised_states=6)
     mat_p_init = cas.DM_eye(6) * np.array(
         [1e-4, 1e-4, 1e-7, 1e-7, 1e-6, 1e-6]
@@ -131,11 +129,11 @@ def configure_stochastic_optimal_control_problem(ocp: OptimalControlProgram, nlp
     ConfigureProblem.configure_dynamics_function(
         ocp,
         nlp,
-        dyn_func=lambda states, controls, parameters, stochastic_variables, nlp, wM, wS: nlp.dynamics_type.dynamic_function(
-            states, controls, parameters, stochastic_variables, nlp, wM, wS, with_gains=False
+        dyn_func=lambda states, controls, parameters, stochastic_variables, nlp, motor_noise, sensory_noise: nlp.dynamics_type.dynamic_function(
+            states, controls, parameters, stochastic_variables, nlp, motor_noise, sensory_noise, with_gains=False
         ),
-        wM=wM,
-        wS=wS,
+        motor_noise=motor_noise,
+        sensory_noise=sensory_noise,
         expand=False,
     )
     return
@@ -148,30 +146,30 @@ def minimize_uncertainty(controllers: list[PenaltyController], key: str) -> cas.
     dt = controllers[0].tf / controllers[0].ns
     out = 0
     for i, ctrl in enumerate(controllers):
-        P_matrix = ctrl.integrated_values["cov"].reshape_to_matrix(
+        cov_matrix = ctrl.integrated_values["cov"].reshape_to_matrix(
             ctrl.integrated_values, ctrl.states.cx.shape[0], ctrl.states.cx.shape[0], Node.START, "cov"
         )
-        P_partial = P_matrix[ctrl.states[key].index, ctrl.states[key].index]
-        out += cas.trace(P_partial) * dt
+        p_partial = cov_matrix[ctrl.states[key].index, ctrl.states[key].index]
+        out += cas.trace(p_partial) * dt
     return out
 
 
-def get_ee(controller: PenaltyController, q, qdot) -> cas.MX:
+def get_ref(controller: PenaltyController, q, qdot) -> cas.MX:
     hand_pos = controller.model.markers(q)[2][:2]
     hand_vel = controller.model.marker_velocities(q, qdot)[2][:2]
     ee = cas.vertcat(hand_pos, hand_vel)
     return ee
 
 
-def ee_equals_ee_ref(controller: PenaltyController) -> cas.MX:
+def hand_equals_ref(controller: PenaltyController) -> cas.MX:
     q = controller.states["q"].cx_start
     qdot = controller.states["qdot"].cx_start
-    ee_ref = controller.stochastic_variables["ee_ref"].cx_start
-    ee = get_ee(controller, q, qdot)
-    return ee - ee_ref
+    ref = controller.stochastic_variables["ref"].cx_start
+    ee = get_ref(controller, q, qdot)
+    return ee - ref
 
 
-def get_p_mat(nlp, node_index, force_field_magnitude, wM_magnitude, wS_magnitude):
+def get_cov_mat(nlp, node_index, force_field_magnitude, motor_noise_magnitude, sensory_noise_magnitude):
     dt = nlp.tf / nlp.ns
 
     nlp.states.node_index = node_index - 1
@@ -186,9 +184,9 @@ def get_p_mat(nlp, node_index, force_field_magnitude, wM_magnitude, wS_magnitude
 
     M_matrix = nlp.stochastic_variables.reshape_to_matrix(nlp.stochastic_variables, nx, nx, Node.START, "m")
 
-    wM = cas.MX.sym("wM", n_tau)
-    wS = cas.MX.sym("wS", n_q + n_qdot)
-    sigma_w = cas.vertcat(wS, wM) * cas.MX_eye(cas.vertcat(wS, wM).shape[0])
+    motor_noise = cas.MX.sym("motor_noise", n_tau)
+    sensory_noise = cas.MX.sym("sensory_noise", n_q + n_qdot)
+    sigma_w = cas.vertcat(sensory_noise, motor_noise) * cas.MX_eye(cas.vertcat(sensory_noise, motor_noise).shape[0])
     cov_sym = cas.MX.sym("cov", nlp.integrated_values.cx_start.shape[0])
     cov_sym_dict = {"cov": cov_sym}
     cov_sym_dict["cov"].cx_start = cov_sym
@@ -200,14 +198,14 @@ def get_p_mat(nlp, node_index, force_field_magnitude, wM_magnitude, wS_magnitude
         nlp.parameters,
         nlp.stochastic_variables.cx_start,
         nlp,
-        wM,
-        wS,
+        motor_noise,
+        sensory_noise,
         force_field_magnitude=force_field_magnitude,
         with_gains=True,
     )
 
-    ddx_dwM = cas.jacobian(dx.dxdt, cas.vertcat(wS, wM))
-    dg_dw = -ddx_dwM * dt
+    ddx_dwm = cas.jacobian(dx.dxdt, cas.vertcat(sensory_noise, motor_noise))
+    dg_dw = -ddx_dwm * dt
     ddx_dx = cas.jacobian(dx.dxdt, nlp.states.cx_start)
     dg_dx = -(ddx_dx * dt / 2 + cas.MX_eye(ddx_dx.shape[0]))
 
@@ -220,8 +218,8 @@ def get_p_mat(nlp, node_index, force_field_magnitude, wM_magnitude, wS_magnitude
             nlp.parameters,
             nlp.stochastic_variables.cx_start,
             cov_sym,
-            wM,
-            wS,
+            motor_noise,
+            sensory_noise,
         ],
         [p_next],
     )(
@@ -230,8 +228,8 @@ def get_p_mat(nlp, node_index, force_field_magnitude, wM_magnitude, wS_magnitude
         nlp.parameters,
         nlp.stochastic_variables.cx_start,
         nlp.integrated_values["cov"].cx_start,  # Should be the right shape to work
-        wM_magnitude,
-        wS_magnitude,
+        motor_noise_magnitude,
+        sensory_noise_magnitude,
     )
     p_vector = nlp.integrated_values.reshape_to_vector(func_eval)
     return p_vector
@@ -244,8 +242,8 @@ def reach_target_consistantly(controllers: list[PenaltyController]) -> cas.MX:
     applies at the END node.
     """
 
-    Q = cas.MX.sym("q_sym", controllers[-1].states["q"].cx_start.shape[0])
-    Qdot = cas.MX.sym("qdot_sym", controllers[-1].states["qdot"].cx_start.shape[0])
+    q_sym = cas.MX.sym("q_sym", controllers[-1].states["q"].cx_start.shape[0])
+    qdot_sym = cas.MX.sym("qdot_sym", controllers[-1].states["qdot"].cx_start.shape[0])
     cov_sym = cas.MX.sym("cov", controllers[-1].integrated_values.cx_start.shape[0])
     cov_sym_dict = {"cov": cov_sym}
     cov_sym_dict["cov"].cx_start = cov_sym
@@ -261,21 +259,21 @@ def reach_target_consistantly(controllers: list[PenaltyController]) -> cas.MX:
         )
     )
 
-    hand_pos = controllers[0].model.markers(Q)[2][:2]
-    hand_vel = controllers[0].model.marker_velocities(Q, Qdot)[2][:2]
+    hand_pos = controllers[0].model.markers(q_sym)[2][:2]
+    hand_vel = controllers[0].model.marker_velocities(q_sym, qdot_sym)[2][:2]
 
-    jac_marker_q = cas.jacobian(hand_pos, Q)
-    jac_marker_qdot = cas.jacobian(hand_vel, cas.vertcat(Q, Qdot))
+    jac_marker_q = cas.jacobian(hand_pos, q_sym)
+    jac_marker_qdot = cas.jacobian(hand_vel, cas.vertcat(q_sym, qdot_sym))
 
-    P_matrix_q = cov_matrix[:2, :2]
-    P_matrix_qdot = cov_matrix[:4, :4]
+    cov_matrix_q = cov_matrix[:2, :2]
+    cov_matrix_qdot = cov_matrix[:4, :4]
 
-    pos_constraint = jac_marker_q @ P_matrix_q @ jac_marker_q.T
-    vel_constraint = jac_marker_qdot @ P_matrix_qdot @ jac_marker_qdot.T
+    pos_constraint = jac_marker_q @ cov_matrix_q @ jac_marker_q.T
+    vel_constraint = jac_marker_qdot @ cov_matrix_qdot @ jac_marker_qdot.T
 
     out = cas.vertcat(pos_constraint[0, 0], pos_constraint[1, 1], vel_constraint[0, 0], vel_constraint[1, 1])
 
-    fun = cas.Function("reach_target_consistantly", [Q, Qdot, cov_sym], [out])
+    fun = cas.Function("reach_target_consistantly", [q_sym, qdot_sym, cov_sym], [out])
     val = fun(
         controllers[-1].states["q"].cx_start,
         controllers[-1].states["qdot"].cx_start,
@@ -286,7 +284,7 @@ def reach_target_consistantly(controllers: list[PenaltyController]) -> cas.MX:
     return val
 
 
-def expected_feedback_effort(controllers: list[PenaltyController], wS_magnitude: cas.DM) -> cas.MX:
+def expected_feedback_effort(controllers: list[PenaltyController], sensory_noise_magnitude: cas.DM) -> cas.MX:
     """
     ...
     """
@@ -295,11 +293,11 @@ def expected_feedback_effort(controllers: list[PenaltyController], wS_magnitude:
     n_qdot = controllers[0].states["qdot"].cx_start.shape[0]
 
     dt = controllers[0].tf / controllers[0].ns
-    sensory_noise_matrix = wS_magnitude * cas.MX_eye(4)
+    sensory_noise_matrix = sensory_noise_magnitude * cas.MX_eye(4)
 
     # create the casadi function to be evaluated
     # Get the symbolic variables
-    ee_ref = controllers[0].stochastic_variables["ee_ref"].cx_start
+    ref = controllers[0].stochastic_variables["ref"].cx_start
     cov_sym = cas.MX.sym("cov", controllers[0].integrated_values.cx_start.shape[0])
     cov_sym_dict = {"cov": cov_sym}
     cov_sym_dict["cov"].cx_start = cov_sym
@@ -316,20 +314,20 @@ def expected_feedback_effort(controllers: list[PenaltyController], wS_magnitude:
     )
 
     k = controllers[0].stochastic_variables["k"].cx_start
-    K_matrix = cas.MX(n_q + n_qdot, n_tau)
+    k_matrix = cas.MX(n_q + n_qdot, n_tau)
     for s0 in range(n_q + n_qdot):
         for s1 in range(n_tau):
-            K_matrix[s0, s1] = k[s0 * n_tau + s1]
-    K_matrix = K_matrix.T
+            k_matrix[s0, s1] = k[s0 * n_tau + s1]
+    k_matrix = k_matrix.T
 
     # Compute the expected effort
     hand_pos = controllers[0].model.markers(controllers[0].states["q"].cx_start)[2][:2]
     hand_vel = controllers[0].model.marker_velocities(
         controllers[0].states["q"].cx_start, controllers[0].states["qdot"].cx_start
     )[2][:2]
-    trace_k_sensor_k = cas.trace(K_matrix @ sensory_noise_matrix @ K_matrix.T)
+    trace_k_sensor_k = cas.trace(k_matrix @ sensory_noise_matrix @ k_matrix.T)
     ee = cas.vertcat(hand_pos, hand_vel)
-    e_fb = K_matrix @ ((ee - ee_ref) + wS_magnitude)
+    e_fb = k_matrix @ ((ee - ref) + sensory_noise_magnitude)
     jac_e_fb_x = cas.jacobian(e_fb, controllers[0].states.cx_start)
     trace_jac_p_jack = cas.trace(jac_e_fb_x @ cov_matrix @ jac_e_fb_x.T)
     expectedEffort_fb_mx = trace_jac_p_jack + trace_k_sensor_k
@@ -349,7 +347,7 @@ def expected_feedback_effort(controllers: list[PenaltyController], wS_magnitude:
 
 
 def zero_acceleration(
-    controller: PenaltyController, wM: np.ndarray, wS: np.ndarray, force_field_magnitude: float
+    controller: PenaltyController, motor_noise: np.ndarray, sensory_noise: np.ndarray, force_field_magnitude: float
 ) -> cas.MX:
     dx = stochastic_forward_dynamics(
         controller.states.cx_start,
@@ -357,8 +355,8 @@ def zero_acceleration(
         controller.parameters.cx_start,
         controller.stochastic_variables.cx_start,
         controller.get_nlp,
-        wM,
-        wS,
+        motor_noise,
+        sensory_noise,
         force_field_magnitude=force_field_magnitude,
         with_gains=False,
     )
@@ -371,13 +369,13 @@ def track_final_marker(controller: PenaltyController) -> cas.MX:
     return ee_pos
 
 
-def leuven_trapezoidal(controllers: list[PenaltyController], force_field_magnitude) -> cas.MX:
+def trapezoidal_integration_continuity_constraint(controllers: list[PenaltyController], force_field_magnitude) -> cas.MX:
     n_q = controllers[0].model.nb_q
     n_qdot = controllers[0].model.nb_qdot
     n_tau = controllers[0].model.nb_tau
 
-    wM = np.zeros((n_tau, 1))
-    wS = np.zeros((n_q + n_qdot, 1))
+    motor_noise = np.zeros((n_tau, 1))
+    sensory_noise = np.zeros((n_q + n_qdot, 1))
     dt = controllers[0].tf / controllers[0].ns
 
     dyn = stochastic_forward_dynamics(
@@ -386,27 +384,27 @@ def leuven_trapezoidal(controllers: list[PenaltyController], force_field_magnitu
         controllers[0].parameters.cx_start,
         controllers[0].stochastic_variables.cx_start,
         controllers[0].get_nlp,
-        wM,
-        wS,
+        motor_noise,
+        sensory_noise,
         force_field_magnitude=force_field_magnitude,
         with_gains=False,
     )
-    dX_i = dyn.dxdt
+    dx_i = dyn.dxdt
 
     if controllers[1].node_index != controllers[0].ns:
-        dX_i_plus = stochastic_forward_dynamics(
+        dx_i_plus = stochastic_forward_dynamics(
             controllers[1].states.cx_start,
             controllers[1].controls.cx_start,
             controllers[1].parameters.cx_start,
             controllers[1].stochastic_variables.cx_start,
             controllers[1].get_nlp,
-            wM,
-            wS,
+            motor_noise,
+            sensory_noise,
             force_field_magnitude=force_field_magnitude,
             with_gains=False,
         ).dxdt
 
-        continuity = controllers[1].states.cx_start - (controllers[0].states.cx_start + (dX_i + dX_i_plus) / 2 * dt)
+        continuity = controllers[1].states.cx_start - (controllers[0].states.cx_start + (dx_i + dx_i_plus) / 2 * dt)
         continuity *= 1e3
     else:
         continuity = 0
@@ -421,8 +419,8 @@ def prepare_socp(
     final_time: float,
     n_shooting: int,
     ee_final_position: np.ndarray,
-    wM_magnitude: cas.DM,
-    wS_magnitude: cas.DM,
+    motor_noise_magnitude: cas.DM,
+    sensory_noise_magnitude: cas.DM,
     force_field_magnitude: float = 0,
     problem_type: str = "CIRCLE",
 ) -> StochasticOptimalControlProgram:
@@ -438,8 +436,10 @@ def prepare_socp(
         The number of shooting points to define int the direct multiple shooting program
     ee_final_position: np.ndarray
         The final position of the end effector
-    ee_initial_position: np.ndarray
-        The initial position of the end effector
+    motor_noise_magnitude: cas.DM
+        The magnitude of the motor noise
+    sensory_noise_magnitude: cas.DM
+        The magnitude of the sensory noise
     force_field_magnitude: float
         The magnitude of the force field
     problem_type: str
@@ -481,14 +481,14 @@ def prepare_socp(
         expected_feedback_effort,
         nodes_phase=[0 for _ in range(n_shooting)],
         nodes=[i for i in range(n_shooting)],
-        wS_magnitude=wS_magnitude,
+        sensory_noise_magnitude=sensory_noise_magnitude,
         weight=1e3 / 2,
         quadratic=False,
     )
 
     # Constraints
     constraints = ConstraintList()
-    constraints.add(ee_equals_ee_ref, node=Node.ALL_SHOOTING)
+    constraints.add(hand_equals_ref, node=Node.ALL_SHOOTING)
     constraints.add(
         ConstraintFcn.TRACK_STATE, key="q", node=Node.START, target=np.array([shoulder_pos_initial, elbow_pos_initial])
     )
@@ -496,8 +496,8 @@ def prepare_socp(
     constraints.add(
         zero_acceleration,
         node=Node.START,
-        wM=np.zeros((n_tau, 1)),
-        wS=np.zeros((n_q + n_qdot, 1)),
+        motor_noise=np.zeros((n_tau, 1)),
+        sensory_noise=np.zeros((n_q + n_qdot, 1)),
         force_field_magnitude=force_field_magnitude,
     )
     constraints.add(track_final_marker, node=Node.PENULTIMATE, target=ee_final_position)
@@ -505,8 +505,8 @@ def prepare_socp(
     constraints.add(
         zero_acceleration,
         node=Node.PENULTIMATE,
-        wM=np.zeros((n_tau, 1)),
-        wS=np.zeros((n_q + n_qdot, 1)),
+        motor_noise=np.zeros((n_tau, 1)),
+        sensory_noise=np.zeros((n_q + n_qdot, 1)),
         force_field_magnitude=force_field_magnitude,
     )  # Not possible sice the control on the last node is NaN
     constraints.add(
@@ -530,26 +530,28 @@ def prepare_socp(
     )
     for i in range(n_shooting):
         multinode_constraints.add(
-            leuven_trapezoidal, nodes_phase=[0, 0], nodes=[i, i + 1], force_field_magnitude=force_field_magnitude
+            trapezoidal_integration_continuity_constraint, 
+            nodes_phase=[0, 0], nodes=[i, i + 1], 
+            force_field_magnitude=force_field_magnitude
         )
 
     # Dynamics
     dynamics = DynamicsList()
     dynamics.add(
         configure_stochastic_optimal_control_problem,
-        dynamic_function=lambda states, controls, parameters, stochastic_variables, nlp, wM, wS, with_gains: stochastic_forward_dynamics(
+        dynamic_function=lambda states, controls, parameters, stochastic_variables, nlp, motor_noise, sensory_noise, with_gains: stochastic_forward_dynamics(
             states,
             controls,
             parameters,
             stochastic_variables,
             nlp,
-            wM,
-            wS,
+            motor_noise,
+            sensory_noise,
             force_field_magnitude=force_field_magnitude,
             with_gains=with_gains,
         ),
-        wM=np.zeros((n_tau, 1)),
-        wS=np.zeros((n_q + n_qdot, 1)),
+        motor_noise=np.zeros((n_tau, 1)),
+        sensory_noise=np.zeros((n_q + n_qdot, 1)),
     )  # expand=False
 
     states_min = np.ones((n_states, n_shooting + 1)) * -cas.inf
@@ -602,10 +604,10 @@ def prepare_socp(
     u_init.add("qdddot", initial_guess=controls_init[:n_q, :], interpolation=InterpolationType.EACH_FRAME)
     u_init.add("tau", initial_guess=controls_init[n_q:, :], interpolation=InterpolationType.EACH_FRAME)
 
-    # TODO: This should probably be done automatically, not defined by the user
+
     s_init = InitialGuessList()
     s_bounds = BoundsList()
-    n_stochastic = n_tau * (n_q + n_qdot) + n_q + n_qdot + n_states * n_states  # K(2x4) + ee_ref(4x1) + M(6x6)
+    n_stochastic = n_tau * (n_q + n_qdot) + n_q + n_qdot + n_states * n_states  # K(2x4) + ref(4x1) + M(6x6)
     stochastic_init = np.zeros((n_stochastic, n_shooting + 1))
     stochastic_min = np.ones((n_stochastic, 3)) * -cas.inf
     stochastic_max = np.ones((n_stochastic, 3)) * cas.inf
@@ -620,16 +622,14 @@ def prepare_socp(
         max_bound=stochastic_max[: n_tau * (n_q + n_qdot), :],
     )
     curent_index += n_tau * (n_q + n_qdot)
-    stochastic_init[curent_index : curent_index + n_q + n_qdot, :] = 0.01  # ee_ref
-    # stochastic_init[curent_index : curent_index + n_q+n_qdot, 0] = np.array([ee_initial_position[0], ee_initial_position[1], 0, 0])  # ee_ref
-    # stochastic_init[curent_index : curent_index + n_q+n_qdot, 1] = np.array([ee_final_position[0], ee_final_position[1], 0, 0])
+    stochastic_init[curent_index : curent_index + n_q + n_qdot, :] = 0.01  # ref
     s_init.add(
-        "ee_ref",
+        "ref",
         initial_guess=stochastic_init[curent_index : curent_index + n_q + n_qdot, :],
         interpolation=InterpolationType.EACH_FRAME,
     )
     s_bounds.add(
-        "ee_ref",
+        "ref",
         min_bound=stochastic_min[curent_index : curent_index + n_q + n_qdot, :],
         max_bound=stochastic_max[curent_index : curent_index + n_q + n_qdot, :],
     )
@@ -647,12 +647,12 @@ def prepare_socp(
     )
 
     integrated_value_functions = {
-        "cov": lambda nlp, node_index: get_p_mat(
+        "cov": lambda nlp, node_index: get_cov_mat(
             nlp,
             node_index,
             force_field_magnitude=force_field_magnitude,
-            wM_magnitude=wM_magnitude,
-            wS_magnitude=wS_magnitude,
+            motor_noise_magnitude=motor_noise_magnitude,
+            sensory_noise_magnitude=sensory_noise_magnitude,
         )
     }
 
@@ -675,15 +675,14 @@ def prepare_socp(
         skip_continuity=True,
         n_threads=1,
         assume_phase_dynamics=False,
-        problem_type=OcpType.SOCP_EXPLICIT(wM_magnitude, wS_magnitude),
+        problem_type=OcpType.SOCP_EXPLICIT(motor_noise_magnitude, sensory_noise_magnitude),
         integrated_value_functions=integrated_value_functions,
     )
 
 
 def main():
-    RUN_OPTIM_FLAG = True  # False
-    PLOT_SOL_FLAG = True
-    VIZUALIZE_SOL_FLAG = True
+    # --- Options --- #
+    vizualize_sol_flag = True
 
     biorbd_model_path = "models/LeuvenArmModel.bioMod"
 
@@ -692,31 +691,27 @@ def main():
 
     # --- Prepare the ocp --- #
     dt = 0.01
-    # final_time = 0.8
-    # n_shooting = int(final_time/dt) + 1
-    # final_time += dt
-    n_shooting = 4
     final_time = 0.8
+    n_shooting = int(final_time/dt) + 1
+    final_time += dt
 
     # --- Noise constants --- #
-    wM_std = 0.05
+    motor_noise_std = 0.05
     wPq_std = 3e-4
     wPqdot_std = 0.0024
 
-    wM_magnitude = cas.DM(np.array([wM_std**2 / dt, wM_std**2 / dt]))
+    motor_noise_magnitude = cas.DM(np.array([motor_noise_std**2 / dt, motor_noise_std**2 / dt]))
     wPq_magnitude = cas.DM(np.array([wPq_std**2 / dt, wPq_std**2 / dt]))
     wPqdot_magnitude = cas.DM(np.array([wPqdot_std**2 / dt, wPqdot_std**2 / dt]))
-    wS_magnitude = cas.vertcat(wPq_magnitude, wPqdot_magnitude)
+    sensory_noise_magnitude = cas.vertcat(wPq_magnitude, wPqdot_magnitude)
 
     # Solver parameters
     solver = Solver.IPOPT(show_online_optim=False)
     solver.set_linear_solver("mumps")
-    # solver.set_linear_solver('ma57')
     solver.set_tol(1e-3)
     solver.set_dual_inf_tol(3e-4)
     solver.set_constr_viol_tol(1e-7)
-    # solver.set_maximum_iterations(10000)
-    solver.set_maximum_iterations(4)
+    solver.set_maximum_iterations(10000)
     solver.set_hessian_approximation("limited-memory")
     solver.set_bound_frac(1e-8)
     solver.set_bound_push(1e-8)
@@ -729,170 +724,54 @@ def main():
         final_time=final_time,
         n_shooting=n_shooting,
         ee_final_position=ee_final_position,
-        wM_magnitude=wM_magnitude,
-        wS_magnitude=wS_magnitude,
+        motor_noise_magnitude=motor_noise_magnitude,
+        sensory_noise_magnitude=sensory_noise_magnitude,
         problem_type=problem_type,
         force_field_magnitude=force_field_magnitude,
     )
 
-    if RUN_OPTIM_FLAG:
-        sol_socp = socp.solve(solver)
-        # print('ici')
+    sol_socp = socp.solve(solver)
 
-        q_sol = sol_socp.states["q"]
-        qdot_sol = sol_socp.states["qdot"]
-        qddot_sol = sol_socp.states["qddot"]
-        qdddot_sol = sol_socp.controls["qdddot"]
-        tau_sol = sol_socp.controls["tau"]
-        k_sol = sol_socp.stochastic_variables["k"]
-        ee_ref_sol = sol_socp.stochastic_variables["ee_ref"]
-        m_sol = sol_socp.stochastic_variables["m"]
-        cov_sol_vect = sol_socp.integrated_values["cov"]
-        cov_sol = np.zeros((6, 6, n_shooting))
-        for i in range(n_shooting):
-            for j in range(6):
-                for k in range(6):
-                    cov_sol[j, k, i] = cov_sol_vect[j * 6 + k, i]
-        stochastic_variables_sol = np.vstack((k_sol, ee_ref_sol, m_sol))
-        data = {
-            "q_sol": q_sol,
-            "qdot_sol": qdot_sol,
-            "qddot_sol": qddot_sol,
-            "qdddot_sol": qdddot_sol,
-            "tau_sol": tau_sol,
-            "k_sol": k_sol,
-            "ee_ref_sol": ee_ref_sol,
-            "m_sol": m_sol,
-            "cov_sol": cov_sol,
-            "stochastic_variables_sol": stochastic_variables_sol,
-        }
+    q_sol = sol_socp.states["q"]
+    qdot_sol = sol_socp.states["qdot"]
+    qddot_sol = sol_socp.states["qddot"]
+    qdddot_sol = sol_socp.controls["qdddot"]
+    tau_sol = sol_socp.controls["tau"]
+    k_sol = sol_socp.stochastic_variables["k"]
+    ref_sol = sol_socp.stochastic_variables["ref"]
+    m_sol = sol_socp.stochastic_variables["m"]
+    cov_sol_vect = sol_socp.integrated_values["cov"]
+    cov_sol = np.zeros((6, 6, n_shooting))
+    for i in range(n_shooting):
+        for j in range(6):
+            for k in range(6):
+                cov_sol[j, k, i] = cov_sol_vect[j * 6 + k, i]
+    stochastic_variables_sol = np.vstack((k_sol, ref_sol, m_sol))
+    data = {
+        "q_sol": q_sol,
+        "qdot_sol": qdot_sol,
+        "qddot_sol": qddot_sol,
+        "qdddot_sol": qdddot_sol,
+        "tau_sol": tau_sol,
+        "k_sol": k_sol,
+        "ref_sol": ref_sol,
+        "m_sol": m_sol,
+        "cov_sol": cov_sol,
+        "stochastic_variables_sol": stochastic_variables_sol,
+    }
 
-        # --- Save the results --- #
-        with open(f"leuvenarm_torque_driven_socp_{problem_type}_forcefield{force_field_magnitude}.pkl", "wb") as file:
-            pickle.dump(data, file)
-    else:
-        with open(f"leuvenarm_torque_driven_socp_{problem_type}_forcefield{force_field_magnitude}.pkl", "rb") as file:
-            data = pickle.load(file)
-        q_sol = data["q_sol"]
-        qdot_sol = data["qdot_sol"]
-        qddot_sol = data["qddot_sol"]
-        qdddot_sol = data["qdddot_sol"]
-        tau_sol = data["tau_sol"]
-        k_sol = data["k_sol"]
-        ee_ref_sol = data["ee_ref_sol"]
-        m_sol = data["m_sol"]
-        # cov_sol = data["cov_sol"]
-        stochastic_variables_sol = np.vstack((k_sol, ee_ref_sol, m_sol))
-
-    if VIZUALIZE_SOL_FLAG:
+    # --- Save the results --- #
+    with open(f"leuvenarm_torque_driven_socp_{problem_type}_forcefield{force_field_magnitude}.pkl", "wb") as file:
+        pickle.dump(data, file)
+    
+    
+    # --- Visualize the results --- #
+    if vizualize_sol_flag:
         import bioviz
 
         b = bioviz.Viz(model_path=biorbd_model_path)
         b.load_movement(q_sol[:, :-1])
         b.exec()
-
-    # --- Plot the results --- #
-    if PLOT_SOL_FLAG:
-        model = biorbd.Model(biorbd_model_path)
-        Q_sym = cas.MX.sym("Q", 2, 1)
-        Qdot_sym = cas.MX.sym("Qdot", 2, 1)
-        hand_pos_fcn = cas.Function("hand_pos", [Q_sym], [model.markers(Q_sym)[2].to_mx()])
-        hand_vel_fcn = cas.Function("hand_vel", [Q_sym, Qdot_sym], [model.markersVelocity(Q_sym, Qdot_sym)[2].to_mx()])
-
-        states = socp.nlp[0].states.cx_start
-        controls = socp.nlp[0].controls.cx_start
-        parameters = socp.nlp[0].parameters.cx_start
-        stochastic_variables = socp.nlp[0].stochastic_variables.cx_start
-        nlp = socp.nlp[0]
-        wM_sym = cas.MX.sym("wM", 2, 1)
-        wS_sym = cas.MX.sym("wS", 4, 1)
-        out = stochastic_forward_dynamics(
-            states,
-            controls,
-            parameters,
-            stochastic_variables,
-            nlp,
-            wM_sym,
-            wS_sym,
-            force_field_magnitude=force_field_magnitude,
-            with_gains=True,
-        )
-        dyn_fun = cas.Function(
-            "dyn_fun", [states, controls, parameters, stochastic_variables, wM_sym, wS_sym], [out.dxdt]
-        )
-
-        fig, axs = plt.subplots(3, 2)
-        n_simulations = 30
-        q_simulated = np.zeros((n_simulations, 2, n_shooting + 1))
-        qdot_simulated = np.zeros((n_simulations, 2, n_shooting + 1))
-        qddot_simulated = np.zeros((n_simulations, 2, n_shooting + 1))
-        hand_pos_simulated = np.zeros((n_simulations, 2, n_shooting + 1))
-        hand_vel_simulated = np.zeros((n_simulations, 2, n_shooting + 1))
-        for i_simulation in range(n_simulations):
-            wM = np.random.normal(0, wM_std, (2, n_shooting + 1))
-            wPq = np.random.normal(0, wPq_std, (2, n_shooting + 1))
-            wPqdot = np.random.normal(0, wPqdot_std, (2, n_shooting + 1))
-            wS = cas.vertcat(wPq, wPqdot)
-            q_simulated[i_simulation, :, 0] = q_sol[:, 0]
-            qdot_simulated[i_simulation, :, 0] = qdot_sol[:, 0]
-            qddot_simulated[i_simulation, :, 0] = qddot_sol[:, 0]
-            for i_node in range(n_shooting):
-                x_prev = cas.vertcat(
-                    q_simulated[i_simulation, :, i_node],
-                    qdot_simulated[i_simulation, :, i_node],
-                    qddot_simulated[i_simulation, :, i_node],
-                )
-                hand_pos_simulated[i_simulation, :, i_node] = np.reshape(hand_pos_fcn(x_prev[:2])[:2], (2,))
-                hand_vel_simulated[i_simulation, :, i_node] = np.reshape(
-                    hand_vel_fcn(x_prev[:2], x_prev[2:4])[:2], (2,)
-                )
-                u = cas.vertcat(qdddot_sol[:, i_node], tau_sol[:, i_node])
-                s = stochastic_variables_sol[:, i_node]
-                k1 = dyn_fun(x_prev, u, [], s, wM[:, i_node], wS[:, i_node])
-                x_next = x_prev + dt * dyn_fun(x_prev + dt / 2 * k1, u, [], s, wM[:, i_node], wS[:, i_node])
-                q_simulated[i_simulation, :, i_node + 1] = np.reshape(x_next[:2], (2,))
-                qdot_simulated[i_simulation, :, i_node + 1] = np.reshape(x_next[2:4], (2,))
-                qddot_simulated[i_simulation, :, i_node + 1] = np.reshape(x_next[4:], (2,))
-            hand_pos_simulated[i_simulation, :, i_node + 1] = np.reshape(hand_pos_fcn(x_next[:2])[:2], (2,))
-            hand_vel_simulated[i_simulation, :, i_node + 1] = np.reshape(
-                hand_vel_fcn(x_next[:2], x_next[2:4])[:2], (2,)
-            )
-            axs[0, 0].plot(
-                hand_pos_simulated[i_simulation, 0, :], hand_pos_simulated[i_simulation, 1, :], color="tab:red"
-            )
-            axs[1, 0].plot(np.linspace(0, final_time, n_shooting + 1), q_simulated[i_simulation, 0, :], color="k")
-            axs[2, 0].plot(np.linspace(0, final_time, n_shooting + 1), q_simulated[i_simulation, 1, :], color="k")
-            axs[0, 1].plot(
-                hand_vel_simulated[i_simulation, 0, :], hand_vel_simulated[i_simulation, 1, :], color="tab:red"
-            )
-            axs[1, 1].plot(np.linspace(0, final_time, n_shooting + 1), qdot_simulated[i_simulation, 0, :], color="k")
-            axs[2, 1].plot(np.linspace(0, final_time, n_shooting + 1), qdot_simulated[i_simulation, 1, :], color="k")
-        hand_pos_without_noise = np.zeros((2, n_shooting + 1))
-        for i_node in range(n_shooting + 1):
-            hand_pos_without_noise[:, i_node] = np.reshape(hand_pos_fcn(q_sol[:, i_node])[:2], (2,))
-        axs[0, 0].plot(hand_pos_without_noise[0, :], hand_pos_without_noise[1, :], color="k")
-        axs[0, 0].plot(ee_initial_position[0], ee_initial_position[1], color="tab:green", marker="o")
-        axs[0, 0].plot(ee_final_position[0], ee_final_position[1], color="tab:red", marker="o")
-        axs[0, 0].set_xlabel("X [m]")
-        axs[0, 0].set_ylabel("Y [m]")
-        axs[0, 0].set_title("Hand position simulated")
-        axs[1, 0].set_xlabel("Time [s]")
-        axs[1, 0].set_ylabel("Shoulder angle [rad]")
-        axs[2, 0].set_xlabel("Time [s]")
-        axs[2, 0].set_ylabel("Elbow angle [rad]")
-        axs[0, 1].set_xlabel("X velocity [m/s]")
-        axs[0, 1].set_ylabel("Y velocity [m/s]")
-        axs[0, 1].set_title("Hand velocity simulated")
-        axs[1, 1].set_xlabel("Time [s]")
-        axs[1, 1].set_ylabel("Shoulder velocity [rad/s]")
-        axs[2, 1].set_xlabel("Time [s]")
-        axs[2, 1].set_ylabel("Elbow velocity [rad/s]")
-        axs[0, 0].axis("equal")
-        plt.tight_layout()
-        plt.savefig("simulated_results.png", dpi=300)
-        plt.show()
-
-    # TODO: integrate to see the error they commit with the trapezoidal
 
 
 if __name__ == "__main__":
